implement forward transfers helper, simplify forward transfers api methods
