+++ sdk/src/main/java/com/horizen/account/api/rpc/types/ForwardTransfersView.java
@@ -1,8 +1,5 @@
 package com.horizen.account.api.rpc.types;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonUnwrapped;
 import com.fasterxml.jackson.annotation.JsonView;
 import com.horizen.evm.utils.Address;
 import com.horizen.serialization.Views;
@@ -10,21 +7,25 @@ import com.horizen.transaction.mainchain.ForwardTransfer;
 import org.web3j.utils.Numeric;
 
 import java.math.BigInteger;
-import java.util.*;
+import java.util.Arrays;
+import java.util.List;
+import java.util.TreeMap;
 
 @JsonView(Views.Default.class)
 public class ForwardTransfersView {
     private final TreeMap<Integer, ForwardTransferData> forwardTransfers = new TreeMap();
 
     public ForwardTransfersView(List<ForwardTransfer> transactions, boolean noPrefix) {
-        if (noPrefix)
-            transactions.forEach(txOutput -> forwardTransfers.put(forwardTransfers.size(),
-                    new ForwardTransferData(Numeric.toHexStringNoPrefix(Arrays.copyOf(txOutput.getFtOutput().propositionBytes(), Address.LENGTH)),
-                    String.valueOf(txOutput.getFtOutput().amount()))));
-        else
-            transactions.forEach(txOutput -> forwardTransfers.put(forwardTransfers.size(),
-                    new ForwardTransferData(Numeric.toHexString(Arrays.copyOf(txOutput.getFtOutput().propositionBytes(), Address.LENGTH)),
-                    Numeric.toHexStringWithPrefix(BigInteger.valueOf(txOutput.getFtOutput().amount())))));
+        if (transactions.size() > 0) {
+            if (noPrefix)
+                transactions.forEach(txOutput -> forwardTransfers.put(forwardTransfers.size(),
+                        new ForwardTransferData(Numeric.toHexStringNoPrefix(Arrays.copyOf(txOutput.getFtOutput().propositionBytes(), Address.LENGTH)),
+                                String.valueOf(txOutput.getFtOutput().amount()))));
+            else
+                transactions.forEach(txOutput -> forwardTransfers.put(forwardTransfers.size(),
+                        new ForwardTransferData(Numeric.toHexString(Arrays.copyOf(txOutput.getFtOutput().propositionBytes(), Address.LENGTH)),
+                                Numeric.toHexStringWithPrefix(BigInteger.valueOf(txOutput.getFtOutput().amount())))));
+        }
     }
 
     public TreeMap<Integer, ForwardTransferData> getForwardTransfers() {
